#20: Update documentation, standardize language, change TXT data files to CSV, rewrite warning/error messages, and add tar files to LFS

+++ lib/gslab_make/write_logs.py
@@ -149,7 +149,7 @@ def log_files_in_output(paths,
             'makelog' : str
                 Path of makelog.
         }
-    depth : int, optional
+    depth : float, optional
         Level of depth when walking through output directory.
 
     Returns
